+++ front-end-tools/src/components/Box.tsx
