+++ front-end-tools/src/components/CCDScanLinks.tsx
